# openapi.yaml
openapi: 3.0.3
info:
  title: Modullar Advancia Healthcare API
  description: Healthcare payment and compliance management platform
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /disputes:
    get:
      summary: List all disputes
      tags: [Disputes]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of disputes
    post:
      summary: Create dispute
      tags: [Disputes]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dispute_number, dispute_reason, amount]
              properties:
                dispute_number:
                  type: string
                transaction_id:
                  type: string
                  format: uuid
                dispute_reason:
                  type: string
                  enum: [fraud, duplicate, product_not_received, service_not_provided, other]
                amount:
                  type: number
                customer_description:
                  type: string
      responses:
        '201':
          description: Dispute created
  /webhooks:
    get:
      summary: List user webhooks
      tags: [Webhooks]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of webhooks
    post:
      summary: Create webhook
      tags: [Webhooks]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, url, secret_key]
              properties:
                name:
                  type: string
                url:
                  type: string
                  format: uri
                secret_key:
                  type: string
                subscribed_events:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Webhook created
...
  /admin/dashboard:
    get:
      summary: Get admin dashboard stats
      tags: [Admin]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Dashboard statistics
  /admin/transactions:
    get:
      summary: List all transactions
      tags: [Admin]
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Transaction list
  /admin/disputes:
    get:
      summary: List all disputes (admin)
      tags: [Admin]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Dispute list
  /admin/disputes/{id}:
    patch:
      summary: Resolve dispute
      tags: [Admin]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution]
              properties:
                resolution:
                  type: string
                  enum: [won, lost, accepted]
                notes:
                  type: string
      responses:
        '200':
          description: Dispute resolved
  /admin/providers:
    get:
      summary: List all providers (admin)
      tags: [Admin]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Provider list
  /admin/audit-log:
    get:
      summary: Get audit log
      tags: [Admin]
      security:
        - BearerAuth: []
      parameters:
        - name: action
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Audit log entries
  /admin/analytics/revenue:
    get:
      summary: Get revenue analytics
      tags: [Admin]
      security:
        - BearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month, year]
      responses:
        '200':
          description: Revenue analytics
...